urlserver.py: Added page_title support
When URLs are encountered, their page title is looked up and added between parens next to the url in the message.
TODO: Store the page title in the database instead and display it separately in the html. Possibly have a setting for it to be turned on/off.
Maybe also have the option to only look it up (once) at display instead of when the message is received.